add memcpy memset functions for C written SCs to prevent wasmer bug
