+++ integrationTests/vm/arwen/testdata/c-api-tests/crypto/cryptoTest.c
@@ -15,32 +15,6 @@ const int Secp256k1SignatureLength = 71;
 const int LIMIT = 10000;
 const int MSG_LENGTH = 100;
 
-void* memset(void *dest, int c, unsigned long n);
-void* memset(void *dest, int c, unsigned long n)
-{
-    int i;
-    char *cdest = (char *)dest;
-    for (i = 0; i < n; i++)
-    {
-        cdest[i] = c;
-    }
-    return dest;
-}
-
-void *memcpy(void *dest, const void *src, unsigned long n);
-void *memcpy(void *dest, const void *src, unsigned long n)
-{
-    char *csrc = (char *)src;
-    char *cdest = (char *)dest;
-
-    for (int i = 0; i < n; i++)
-    {
-        cdest[i] = csrc[i];
-    }
-
-    return dest;
-}
-
 void init() 
 {
